Added sessionStorage for available table and modal for reserved button.
The sessionStorage store the tableID of the selected available table. The value stored is transfered to the createOrder form to be submitted with the order form to database. Also added a modal that will appear if reserved table button is pressed. Currently the modal contain no information as the customer reserved table in database is not up yet.